requirements.txt
numpy==1.24.1
Pillow==9.4.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
